rename counterparty_wallet -> uniparty
