+++ lib/app_router.dart
@@ -1,7 +1,7 @@
-import 'package:counterparty_wallet/start_page.dart';
-import 'package:counterparty_wallet/secure_utils/secure_storage.dart';
-import 'package:counterparty_wallet/components/wallet_pages/wallet.dart';
 import 'package:go_router/go_router.dart';
+import 'package:uniparty/components/wallet_pages/wallet.dart';
+import 'package:uniparty/secure_utils/secure_storage.dart';
+import 'package:uniparty/start_page.dart';
 
 class AppRouter {
   GoRouter get router => _goRouter;
