+++ lib/components/wallet_creation_pages/create_wallet_button.dart
@@ -1,6 +1,6 @@
-import 'package:counterparty_wallet/secure_utils/bip39.dart';
-import 'package:counterparty_wallet/components/wallet_creation_pages/create_wallet_dialog.dart';
 import 'package:flutter/material.dart';
+import 'package:uniparty/components/wallet_creation_pages/create_wallet_dialog.dart';
+import 'package:uniparty/secure_utils/bip39.dart';
 
 class CreateWalletButton extends StatelessWidget {
   const CreateWalletButton({super.key});
@@ -13,8 +13,7 @@ class CreateWalletButton extends StatelessWidget {
 
         showDialog<String>(
             context: context,
-            builder: (BuildContext context) =>
-                CreateWalletDialog(mnemonic: mnemonic));
+            builder: (BuildContext context) => CreateWalletDialog(mnemonic: mnemonic));
       },
       style: _buttonStyle(),
       child: const Text('Create Wallet'),
@@ -24,7 +23,6 @@ class CreateWalletButton extends StatelessWidget {
   ButtonStyle _buttonStyle() {
     return ButtonStyle(
         fixedSize: MaterialStateProperty.all<Size?>(const Size(200.0, 75.0)),
-        textStyle: MaterialStateProperty.all<TextStyle?>(
-            const TextStyle(fontSize: 15)));
+        textStyle: MaterialStateProperty.all<TextStyle?>(const TextStyle(fontSize: 15)));
   }
 }
