+++ lib/main.dart
@@ -1,6 +1,6 @@
-import 'package:counterparty_wallet/app_router.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
+import 'package:uniparty/app_router.dart';
 
 void main() async {
   await dotenv.load();
