+++ lib/wallet_recovery/counterwallet_recovery.dart
@@ -1,12 +1,12 @@
 import 'dart:typed_data';
 
 import 'package:convert/convert.dart';
-import 'package:counterparty_wallet/secure_utils/bech32.dart';
-import 'package:counterparty_wallet/secure_utils/bip32.dart';
-import 'package:counterparty_wallet/secure_utils/legacy_address.dart';
-import 'package:counterparty_wallet/secure_utils/mnemonic_js.dart';
-import 'package:counterparty_wallet/secure_utils/models/key_pair.dart';
-import 'package:counterparty_wallet/secure_utils/models/wallet_info.dart';
+import 'package:uniparty/secure_utils/bech32.dart';
+import 'package:uniparty/secure_utils/bip32.dart';
+import 'package:uniparty/secure_utils/legacy_address.dart';
+import 'package:uniparty/secure_utils/mnemonic_js.dart';
+import 'package:uniparty/secure_utils/models/key_pair.dart';
+import 'package:uniparty/secure_utils/models/wallet_info.dart';
 
 class CounterwalletRecovery {
   final mnemonicJs = MnemonicJs();
