+++ macos/Runner/Configs/AppInfo.xcconfig
@@ -5,10 +5,10 @@
 // 'flutter create' template.
 
 // The application's name. By default this is also the title of the Flutter window.
-PRODUCT_NAME = counterparty_wallet
+PRODUCT_NAME = uniparty
 
 // The application's bundle identifier
-PRODUCT_BUNDLE_IDENTIFIER = com.example.counterpartyWallet
+PRODUCT_BUNDLE_IDENTIFIER = com.example.uniparty
 
 // The copyright displayed in application information
 PRODUCT_COPYRIGHT = Copyright © 2024 com.example. All rights reserved.
