+++ pubspec.yaml
@@ -1,5 +1,5 @@
-name: counterparty_wallet
-description: "Counterparty wallet"
+name: uniparty
+description: "A wallet for XCP"
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
