+++ test/bip32/bip32_mainnet_test.dart
@@ -1,11 +1,11 @@
 import 'dart:typed_data';
 
 import 'package:convert/convert.dart';
-import 'package:counterparty_wallet/secure_utils/bip32.dart';
-import 'package:counterparty_wallet/secure_utils/bip39.dart';
-import 'package:counterparty_wallet/secure_utils/models/key_pair.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:test/test.dart';
+import 'package:uniparty/secure_utils/bip32.dart';
+import 'package:uniparty/secure_utils/bip39.dart';
+import 'package:uniparty/secure_utils/models/key_pair.dart';
 
 void main() async {
   await dotenv.load();
