+++ test/wallet_recovery/uniparty_recovery/uniparty_recovery_testnet_test.dart
@@ -1,7 +1,7 @@
-import 'package:counterparty_wallet/secure_utils/models/wallet_info.dart';
-import 'package:counterparty_wallet/wallet_recovery/uniparty_recovery.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:test/test.dart';
+import 'package:uniparty/secure_utils/models/wallet_info.dart';
+import 'package:uniparty/wallet_recovery/uniparty_recovery.dart';
 
 void main() async {
   await dotenv.load();
