+++ web/index.html
@@ -23,13 +23,13 @@
   <!-- iOS meta tags & icons -->
   <meta name="apple-mobile-web-app-capable" content="yes">
   <meta name="apple-mobile-web-app-status-bar-style" content="black">
-  <meta name="apple-mobile-web-app-title" content="counterparty_wallet">
+  <meta name="apple-mobile-web-app-title" content="uniparty">
   <link rel="apple-touch-icon" href="icons/Icon-192.png">
 
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>counterparty_wallet</title>
+  <title>uniparty</title>
   <link rel="manifest" href="manifest.json">
 
   <script>
